+++ src/rt_ted.c
@@ -3600,8 +3600,8 @@ void LinkElevatorDiskGroups(void)
 =================
 */
 
-void LinkActor(objtype* ob, int tilex, int tiley, void (*action)(long),
-			   void (*swapaction)(long))
+void LinkActor(objtype* ob, int tilex, int tiley, void (*action)(intptr_t),
+			   void (*swapaction)(intptr_t))
 {
 	word touchx, touchy;
 	int clockx, clocky;
@@ -3616,7 +3616,7 @@ void LinkActor(objtype* ob, int tilex, int tiley, void (*action)(long),
 		if ((clockx == tilex) && (clocky == tiley))
 		{
 			clocklinked = 1;
-			ClockLink(EnableObject, DisableObject, (long)ob, k);
+			ClockLink(EnableObject, DisableObject, (intptr_t)ob, k);
 		}
 	}
 
@@ -3642,10 +3642,10 @@ void LinkActor(objtype* ob, int tilex, int tiley, void (*action)(long),
 			}
 
 			if (tswitch && (tswitch->flags & FL_ON))
-				Link_To_Touchplate(touchx, touchy, swapaction, action, (long)ob,
+				Link_To_Touchplate(touchx, touchy, swapaction, action, (intptr_t)ob,
 								   0);
 			else
-				Link_To_Touchplate(touchx, touchy, action, swapaction, (long)ob,
+				Link_To_Touchplate(touchx, touchy, action, swapaction, (intptr_t)ob,
 								   0);
 			if (ob->obclass == gasgrateobj)
 			{
@@ -3675,7 +3675,7 @@ void SetupInanimateActors(void)
 {
 	int i, j, linked;
 	word *map, tile;
-	void (*action)(long), (*swapaction)(long);
+	void (*action)(intptr_t), (*swapaction)(intptr_t);
 
 	map = mapplanes[1];
 
@@ -4051,7 +4051,7 @@ void SetupLights(void)
 							{
 								Link_To_Touchplate(
 									touchx, touchy, ActivateLight,
-									DeactivateLight, (long)(sprites[i][j]), 0);
+									DeactivateLight, (intptr_t)(sprites[i][j]), 0);
 								sprites[i][j]->linked_to =
 									touchindices[touchx][touchy] - 1;
 							}
@@ -4066,7 +4066,7 @@ void SetupLights(void)
 							{
 								Link_To_Touchplate(
 									touchx, touchy, DeactivateLight,
-									ActivateLight, (long)(sprites[i][j]), 0);
+									ActivateLight, (intptr_t)(sprites[i][j]), 0);
 								sprites[i][j]->linked_to =
 									touchindices[touchx][touchy] - 1;
 							}
@@ -4083,7 +4083,7 @@ void SetupLights(void)
 						{
 							Link_To_Touchplate(touchx, touchy, DeactivateLight,
 											   ActivateLight,
-											   (long)(sprites[i][j]), 0);
+											   (intptr_t)(sprites[i][j]), 0);
 							sprites[i][j]->linked_to =
 								touchindices[touchx][touchy] - 1;
 						}
